+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added test for randomElementFromSlice
 - added tests for urlsHaveSameDomain
 - added link to proxy-list in readme
+- added tests for formatDuration
 
 ## 0.2.1 - 2021.08.12